// src/app/models/employee.model.ts
export interface Employee {
    id: string;
    firstName: string;
    lastName: string;
    cnp: string;
    drivingLicenseExDate: number;
    drivingLicenseCategories: string[];
    email: string;
    phone: string;
    jobDepartment: string;
    emergencyContactName: string;
    emergencyContactPhoneNumber: string;
  }